refactor move all hydrology module names to lower case
