+++ pywatershed/hydrology/__init__.py
@@ -1,8 +1,8 @@
 from .prms_canopy import PRMSCanopy
-from .PRMSChannel import PRMSChannel
+from .prms_channel import PRMSChannel
 
 # from PRMSEt import PRMSEt
-from .PRMSGroundwater import PRMSGroundwater
-from .PRMSRunoff import PRMSRunoff
-from .PRMSSnow import PRMSSnow
-from .PRMSSoilzone import PRMSSoilzone
+from .prms_groundwater import PRMSGroundwater
+from .prms_runoff import PRMSRunoff
+from .prms_snow import PRMSSnow
+from .prms_soilzone import PRMSSoilzone
